==> notebooks/01_Summary_hand_hygiene.ipynb <==
SUMMARY OF MY PUBLICATION ON HAND HYGIENE ANALYSIS IN CRITICAL AND SEMI-CRITICAL CARE AREAS OF TERTIARY CARE HOSPITAL

link: https://files01.core.ac.uk/download/482242137.pdf

This study was conducted on Hand Hygiene Compliance in Critical and Semi-Critical care areas of a Tertiary care hospital during 2017 which later got published in 2018 in “Innovare Journal of Health Sciences, Vol 6, Issue 1”. 

*****************TARGET AREAS AND GROUPS********************************

As per the recommendations and standards of WHO, the direct observation was conducted on healthcare workers in 2 types of Care Units of a 350 bedded Tertiary Care Hospital.

The two types of Care Units were:
1.	CRITICAL CARE UNITS : 
a.	MICU1: Medical Intensive Care Unit 1
b.	MICU2: Medical Intensive Care Unit 2
c.	NeuroICU: Neuro-Intensive Care Unit
d.	CCU: Coronary Care Unit

2.	SEMI-CRITICAL CARE UNITS:
a.	EMR: Emergency
b.	SR: Surgical Recovery
c.	CR: Cardiac Recovery
d.	HDU: High Dependency Unit
e.	Dialysis Unit


The targeted group of Health Care workers was:  
1.	Doctors
2.	Nurses
3.	General Duty Assistant (GDA) and Housekeeper (HK)
4.	Other (Technician, Dietician, Physiotherapist) 


*******************MONITORING / SURVEILLANCE****************************

The WHO recommended observation form was used to record the Compliance during 5 Moments of Hand Hygiene. The adherence to 7 Steps of Hand Cleaning was also monitored along with the Preference of type of Hand Hygiene (Hand Rub / Hand Wash).
The 5 moments of Hand Hygiene as per WHO Observation Form:
1.	bef.pat : Before touching a patient
2.	bef.asept : Before clean/aseptic procedure
3.	aft.b.f : After body fluid exposure risk
4.	aft.pat : After touching a patient
5.	aft.p.surr : After touching patient surroundings


*********************INTERPRETATIONS******************************

After the collection of data, the data was descriptively analysed to obtain results:
1.	Overall Compliance
2.	Compliance during 5 Moments of Hand hygiene
3.	Compliance with 7 Steps of Hand Hygiene 
4.	Preference of Hand-Rub v/s Hand-Wash among healthcare workers
5.	Critical and Semi-Critical Compliance
6.	Area-wise Compliance in Critical and Semi-Critical Care Unit
7.	Moment-wise Compliance


******************COMPUTATIONAL ANALYSIS***************************

This project is used to reanalyze or recalulate the descriptive statistics used in this study to obtain computational analysis of same results along with some other statistical analysis such as chi-square test, z-test. Line charts and tables are generated to show these results.
