==> notebooks/02_Data_preparation.ipynb <==
This Notebook is used to store data obtained from the publication in form of dictionaries and convert the stored data in to CSV files for further Descriptive, Visual and Statistical Analysis. 

2.1 IMPORTING LIBRARIES

import pandas as pd

2.2 CREATING DICTIONARIES

# Dictionary 1: shows 
# 1. Total Opportunities occured during 5 Moments of Hand Hygiene
# 2. Number of Actions performed during 5 Moments of Hand Hygiene
# 3. 7 Steps of Hand cleaning followed during 5 Moments of Hand Hygiene
compliance = {
    "Total_Opportunities" : [1198], 
    "Action_5_Moments" : [483], 
    "Action_7_Steps" : [303]
}




# Dictionary 2 : shows
# 1. Categories of Health Care Workers
# 2. Total Opportunities occured during 5 Moments of Hand Hygiene of each healthcare worker group
# 3. Number of Actions performed by each Health Care Worker group during 5 Moments of Hand Hygiene
# 4. 7 Steps of Hand cleaning followed by each Health Care Worker group during 5 Moments of Hand Hygiene
moment_step_compliance = {
    "HCW" : ["Overall", "Doctors", "Nurses", "GDA/HK", "Other"],
    "Opportunities" : [1198, 236, 509, 172, 281],
    "Action_5_Moments" : [483, 64, 246, 73, 100],
    "Action_7_Steps" : [303, 36, 168, 45, 54]
}




# Dictionary 3 : shows
# 1. Categories of Health Care Workers
# 2. Number of Actions performed by each Health Care Worker group during 5 Moments of Hand Hygiene
# 3. Preference of Hand Rub among Health Care Workers as per their categories for cleaning hands
# 4. Preference of Hand Wash among Health Care Workers as per their categories for cleaning hands
preference = {
    "HCW" : ["overall", "Doctors", "Nurses", "GDA/HK", "Other"],
    "Action_5_Moments" : [483, 64, 246, 73, 100],
    "Hand_Rub" : [334, 52, 178, 37, 63],
    "Hand_Wash" : [157, 11, 67, 40, 37]
    }



# Dictionary 4 : shows
# 1. Categories of Health Care Workers
# 2. In Critical Care Areas - Total Opportunities occured during 5 Moments of Hand Hygiene for different categories of healthcare worker 
# 3. Number of Actions performed during 5 Moments of Hand Hygiene in Critical Care Areas by different categories of healthcare worker 
# 4. In Semi-Critical Care Areas - Total Opportunities occured during 5 Moments of Hand Hygiene for different categories of healthcare worker  
# 5. Number of Actions performed during 5 Moments of Hand Hygiene in Semi-Critical Care Areas by different categories of healthcare worker 
care_area_compliance = {
    "HCW" : ["Overall","Doctors", "Nurses", "GDA/HK", "Other"],
    "C_Opp" : [527, 114, 242, 76, 95],
    "C_Actions" : [229, 32, 125, 41, 31],
    "S_Opp" : [661, 122, 257, 96, 186],
    "S_Actions" : [244, 32, 111, 32, 69]
    }



# Dictionary 5 : shows 
# 1. Critical and Semi-Critical Care Areas
# 2. Overall Compliance in percentage in different Care Area
# 3. Compliance among Doctors in percentage in different Care Areas
# 4. Compliance among Nurses in percentage in different Care Areas
# 5. Compliance among General Duty Assistant (GDA)/ Housekeepers (HK) in percentage in different Care Areas
# 6. Compliance among Others (Technician, Dietician, Physiotherapy) in percentage in different Care Areas
areawise_compliance = {
    "Areas" : ["MICU1", "MICU2", "NeuroICU","CCU", "HDUN", "CS", "SR", "EMR", "CR", "Dialysis" ],
    "Overall (%)" : [50.98, 47.3, 41.18, 30.47, 46.38, 34.28, 40.66, 24.62, 40.37, 45.94],
    "Doctors (%)" : [43.85, 16.6, 5.8, 13.6, 20, 100, 23.52, 25, 25, 0],
    "Nurses (%)": [54.87, 65.90, 57.5, 36.84, 70, 14.28, 50, 30, 50, 0],
    "GDA/HK (%)" : [68.57, 45.45, 50, 28.57, 50, 50, 44.44, 22.22, 23.07, 5.88],
    "Other (%)" : [33.33, 35, 34.48, 25, 52.63, 33.33, 34.78, 15.90, 32.69, 61.53]
}



# Dictionary 6 : shows 
# 1. 5 Moments of hand hygiene
# 2. Number of Opportunities occured as per the Moment of Hand hygiene
# 3. Number of Actions performed during Opportunities occured as per the Moment of Hand hygiene
momentwise = {
    "Moments" : [
        "Before Patient Contact",
        "Before Asceptic Procedure",
        "After Body Fluid Exposure",
        "After Patient Contact",
        "After Touching Patient Surroundings"
    ],
    "Opportunities" : [229, 44, 50, 325, 550],
    "Action" : [69, 15, 16, 173, 210]
}

2.3 CONVERTING DICTIONARIES INTO DATAFRAMES

# USING DATAFRAMES TO CONVERT DICTIONARIES IN TO TABLES

compliance = pd.DataFrame(compliance)
moment_step_compliance = pd.DataFrame(moment_step_compliance)
preference = pd.DataFrame(preference)
care_area_compliance = pd.DataFrame(care_area_compliance)
areawise_compliance = pd.DataFrame(areawise_compliance)
momentwise = pd.DataFrame(momentwise)

2.4 EXPORTING INTO CSV FILES

# USING to_csv() FUNCTION  {to export data into structured file format for storage and sharing} 

compliance.to_csv("overall_compliance_HH.csv", index = False)
moment_step_compliance.to_csv("Moments_Steps_compliance_HH.csv", index = False)
preference.to_csv("Preference_HR_HW.csv", index = False)
care_area_compliance.to_csv("Critical_Semicritical_Area_Compliance.csv", index = False)
areawise_compliance.to_csv("Area_Wise_Compliance.csv", index = False)
momentwise.to_csv("Moment_Wise_Compliance.csv", index = False)

2.5 PRINTING RESULTS/TABLES

# Table 1 : shows
# The Number of Actions performed during 5 Moments and 7 Steps of Hand Hygiene and Total Opportunities occured during these 5 Moments

compliance 

# Table 2 : shows
# The Number of Actions performed during 5 Moments and 7 Steps of Hand hygiene as per different categories of Health Care Workers

moment_step_compliance

# Table 3 : shows
# The Number of Actions performed during 5 Moments and preference of Hand Rub and Hand Hygiene as per different categories of Health Care Workers

preference

# Table 4 : shows
# The Opportunities and Actions occured in Critical and Semi-Critical Care Areas as per different categories of Health Care Workers

care_area_compliance

# Table 5 : shows
# The Overall and Healthcare Worker Wise Percentage of Actions performed during 5 Moments of Hand Hygiene in 9 different Care units

areawise_compliance

# Table 6 : shows
# The Opportunities and the Actions occured as per 5 Moments of Hand Hygiene

momentwise

2.6 LOADING DATA

# USING read.csv() Functions {to read data from CSV file and load the data in to pandas DataFrame}

compliance = pd.read_csv("overall_compliance_HH.csv")
moment_step_compliance = pd.read_csv("Moments_Steps_compliance_HH.csv")
preference = pd.read_csv("Preference_HR_HW.csv")
care_area_compliance = pd.read_csv("Critical_Semicritical_Area_Compliance.csv")
areawise_compliance = pd.read_csv("Area_Wise_Compliance.csv")
momentwise = pd.read_csv("Moment_Wise_Compliance.csv")

2.7 CHECKING EMPTY VALUES

print(compliance.isnull().sum())
print(moment_step_compliance.isnull().sum())
print(preference.isnull().sum())
print(care_area_compliance.isnull().sum())
print(areawise_compliance.isnull().sum())
print(momentwise.isnull().sum())

2.8 PRINTING SUMMARY OF DATA FRAME

# USING .infor() FUNCTION {to print summary/ Information of DataFrame}

print(compliance.info())
print(moment_step_compliance.info())
print(preference.info())
print(care_area_compliance.info())
print(areawise_compliance.info())
print(momentwise.info())

2.9 DATA PREPARATION DONE!!
