==> notebooks/03_Descriptive_Analysis.ipynb <==
This Notebook is used to Compute Compliances with the help of Descriptive Statistics
1. Overall Compliance during 5 Moments with 7 Steps of Hand Hygiene
2. Compliance among Health Care Workers during 5 Moments with 7 Steps of Hand Hygiene
3. Overall preference of Hand-Rub v/s Hand-Wash
4. Overall Compliance of HealthCare Workers in Critical and Semi-Critical Care Area
5. Area Wise Compliance in Critical and Semi-Critical Care Area
6. Compliance as per Moments of Hand Hygiene

The formula used to analyse most of the Complaince (Recomended by WHO):

Compliance (%) = (Actions / Opportunities) * 100


3.1 IMPORTING LIBRARIES

import pandas as pd

3.2 IMPORTING THE CSV FILES FOR DESCRIPTIVE ANALYSIS

# USING read.csv() Functions {to read data from CSV file and load the data in to pandas DataFrame}

compliance = pd.read_csv("overall_compliance_HH.csv")
moment_step_compliance = pd.read_csv("Moments_Steps_compliance_HH.csv")
preference = pd.read_csv("Preference_HR_HW.csv")
care_area_compliance = pd.read_csv("Critical_Semicritical_Area_Compliance.csv")
areawise_compliance = pd.read_csv("Area_Wise_Compliance.csv")
momentwise = pd.read_csv("Moment_Wise_Compliance.csv")

3.3 CALCULATE OVERALL COMPLIANCE (%) DURING 5 MOMENTS OF HAND HYGIENE WITH 7 STEPS OF HAND HYGIENE

# Formula to Calculate Overall Compliance = (Actions / Opportunities) * 100
compliance["Moments(%)"] = (compliance["Action_5_Moments"] / compliance["Total_Opportunities"]) * 100 

compliance["Steps(%)"] = (compliance["Action_7_Steps"] / compliance["Action_5_Moments"]) * 100

# Rounding Off the decimal value up to 2
compliance["Moments(%)"] = compliance["Moments(%)"].round(2)
compliance["Steps(%)"] = compliance["Steps(%)"].round(2)

# Print overall compliance
print(compliance)

3.4 CALCULATE COMPLIANCE (%) AMONG HEALTH CARE WORKERS DURING 5 MOMENTS WITH 7 STEPS OF HAND HYGIENE


# Formula to Calculate Compliance = (Actions / Opportunities) * 100
moment_step_compliance["Moments(%)"] = (moment_step_compliance["Action_5_Moments"] / moment_step_compliance["Opportunities"]) * 100 
moment_step_compliance["Steps(%)"] = (moment_step_compliance["Action_7_Steps"] / moment_step_compliance["Action_5_Moments"]) * 100


# Rounding Off the decimal value up to 2
moment_step_compliance["Moments(%)"] = moment_step_compliance["Moments(%)"].round(2)
moment_step_compliance["Steps(%)"] = moment_step_compliance["Steps(%)"].round(2)

# Print compliance in Moments and Steps
print(moment_step_compliance)


3.5 CALCULATE OVERALL PREFERENCE (%) OF HAND RUB V/S HAND WASH


preference ["HR (%)"] = (preference["Hand_Rub"] / preference["Action_5_Moments"] ) * 100
preference["HW (%)"] = (preference["Hand_Wash"] / preference["Action_5_Moments"] ) * 100

# Rounding Off the decimal value up to 2
preference["HR (%)"] = preference["HR (%)"].round(2)
preference["HW (%)"] = preference["HW (%)"].round(2)

# Print percentage of compliance in Moments and Steps
print(preference)


3.6 CALCULATE OVERALL COMPLIANCE (%) OF HEALTH CARE WORKERS IN CRITICAL AND SEMI-CRITICAL AREA

# C = Critical care area
# SC = Semi-Critical care area

care_area_compliance ["C (%)"] = (care_area_compliance["C_Actions"] / care_area_compliance["C_Opp"]) * 100
care_area_compliance ["SC (%)"] = (care_area_compliance["S_Actions"] / care_area_compliance["S_Opp"]) * 100

# Rounding Off the decimal value up to 2
care_area_compliance["C (%)"] = care_area_compliance["C (%)"].round(2)
care_area_compliance["SC (%)"] = care_area_compliance["SC (%)"].round(2)

# Print compliance in critical and semicritical care areas
print(care_area_compliance)

3.7 CALCULATE AREA WISE COMPLIANCE (%) IN CRITICAL AND SEMI CRITICAL CARE AREAS

areawise_compliance

3.8 CALCULATE COMPLIANCE(%) AS PER MOMENTS OF HAND HYGIENE

momentwise ["Compliance (%)"] = (momentwise["Action"] / momentwise["Opportunities"] * 100)

# Rounding Off the decimal value up to 2
momentwise ["Compliance (%)"] = momentwise ["Compliance (%)"].round(2)

# Print the moment wise compliance percentage
print(momentwise)

3.9 SAVING THE PROCESSED DATA IN CSV

# USING to_csv() FUNCTION  {to export data into structured file format for storage and sharing} 

compliance.to_csv("overall_compliance_HH.csv", index = False)
moment_step_compliance.to_csv("Moments_Steps_compliance_HH.csv", index = False)
preference.to_csv("Preference_HR_HW.csv", index = False)
care_area_compliance.to_csv("Critical_Semicritical_Area_Compliance.csv", index = False)
areawise_compliance.to_csv("Area_Wise_Compliance.csv", index = False)
momentwise.to_csv("Moment_Wise_Compliance.csv", index = False)

3.10 DESCRIPTIVE STATISTICS DONE!!
