==> notebooks/04_Comparitive_Analysus.ipynb <==
This Notebook does comparitive analysis on the data obtained from the descriptive analysis
1. Moment and steps wise complaince comparison among healthcare workers 
2. Preference Handrub v/s Handwash among HCW'S
3. Complaince comparison in Critical and Semicritical care areas
4. Area wise complaince analysis
5. Compliance comparison during 5 moments of hand hygiene

4.1 IMPORTING LIBRARIES

import pandas as pd

4.2 IMPORTING CSV FILES

# USING read.csv() Functions {to read data from CSV file and load the data in to pandas DataFrame}

compliance = pd.read_csv("overall_compliance_HH.csv")
moment_step_compliance = pd.read_csv("Moments_Steps_compliance_HH.csv")
preference = pd.read_csv("Preference_HR_HW.csv")
care_area_compliance = pd.read_csv("Critical_Semicritical_Area_Compliance.csv")
areawise_compliance = pd.read_csv("Area_Wise_Compliance.csv")
momentwise = pd.read_csv("Moment_Wise_Compliance.csv")

4.3 COMPARING MOMENT AND STEP WISE COMPLAINCE AMONG HCW'S

moment_step_compliance[["HCW", "Moments(%)", "Steps(%)"]]

#MOMENT AND STEP WISE HIGHEST AND LOWEST COMPLIANCE AMONG HEALTH CARE WORKERS 

max_moment = moment_step_compliance.loc[moment_step_compliance["Moments(%)"].idxmax()]
max_steps = moment_step_compliance.loc[moment_step_compliance["Steps(%)"].idxmax()]
min_moments = moment_step_compliance.loc[moment_step_compliance["Moments(%)"].idxmin()]
min_steps = moment_step_compliance.loc[moment_step_compliance["Steps(%)"].idxmin()]

print(max_moment) # Nurses - Highest 5 moment compliance
print(max_steps) # Nurses - Highest 7 steps compliance
print(min_moments) # Doctors - Lowest 5 moment compliance
print(min_steps) # Other - Lowest 7 steps compliance

4.4 COMPARING PREFERENCE HAND-RUB V/S HAND-WASH AMONG HCW'S

preference[["HCW", "HR (%)", "HW (%)"]]

#PREFERENCE OF HAND RUB V/S HAND WASH (HIGHEST & LOWEST VALUE)

HR_highest = preference.loc[preference["HR (%)"].idxmax()]
HR_lowest = preference.loc[preference["HW (%)"].idxmax()]
HW_highest = preference.loc[preference["HR (%)"].idxmin()]
HW_lowest = preference.loc[preference["HW (%)"].idxmin()]

print(HR_highest) # Doctors prefers Hand rub more
print(HR_lowest) # GDA/HK prefers Hand rub less
print(HW_highest) # GDA/HK prefers Hand wash more
print(HW_lowest) # Doctors prefers Hand wash less

# Doctors prefers to use hand rub than that of hand wash 
# while GDA/HK prefers to wash their hands

4.5 COMPARING COMPLIANCE OF HCW'S IN CRITICAL AND SEMI CRITICAL CARE AREA

care_area_compliance[["HCW", "C (%)", "SC (%)"]]

#LOWEST AND HIGHEST COMPLICANCE OF HCW'S IN CRITICAL AND SEMICRITICAL CARE UNIT

max_critical_compliance = care_area_compliance.loc[care_area_compliance["C (%)"].idxmax()]
max_semicritical_compliance = care_area_compliance.loc[care_area_compliance["SC (%)"].idxmax()]
min_critical_compliance = care_area_compliance.loc[care_area_compliance["C (%)"].idxmin()]
min_semicritical_compliance = care_area_compliance.loc[care_area_compliance["SC (%)"].idxmin()]

print(max_critical_compliance) # GDA/HK
print(max_semicritical_compliance) # Nurses
print(min_critical_compliance) # Doctors
print(min_semicritical_compliance) # Doctors

# Doctors showed the least compliance while GDA/HK indicates the maximum complaince in critical while nurses in semi critical care areas

4.6 COMPARING AREA WISE COMPLIANCE OF HCW'S

areawise_compliance[["Areas", "Overall (%)", "Doctors (%)","Nurses (%)","GDA/HK (%)","Other (%)"]]

# Minimum and Maximum Complaince in Critical and Semicritical care areas

max_overall = areawise_compliance.loc[areawise_compliance["Overall (%)"].idxmax()]
max_doctors = areawise_compliance.loc[areawise_compliance["Doctors (%)"].idxmax()]
max_nurse = areawise_compliance.loc[areawise_compliance["Nurses (%)"].idxmax()]
max_GDA = areawise_compliance.loc[areawise_compliance["GDA/HK (%)"].idxmax()]
max_other = areawise_compliance.loc[areawise_compliance["Other (%)"].idxmax()]


print(max_overall) # MICU1
print(max_doctors) # CS
print(max_nurse) # HDUN
print(max_GDA) # MICU1
print(max_other) # DIALYSIS

# the critical care units such as MICU1, CS, HDUN showed high complaince percentage
# Other healthcare workers shows high complaince in dialysis

# Minimum and Maximum Complaince in Critical and Semicritical areas

min_overall = areawise_compliance.loc[areawise_compliance["Overall (%)"].idxmin()]
min_doctors = areawise_compliance.loc[areawise_compliance["Doctors (%)"].idxmin()]
min_nurse= areawise_compliance.loc[areawise_compliance["Nurses (%)"].idxmin()]
min_GDA = areawise_compliance.loc[areawise_compliance["GDA/HK (%)"].idxmin()]
min_other = areawise_compliance.loc[areawise_compliance["Other (%)"].idxmin()]

print(min_overall) # EMR
print(min_doctors) # DIALYSIS
print(min_nurse) # Dialysis
print(min_GDA) # DIALYSIS
print(min_other) # EMR

# Emergency and Dialysis unit shows the least complaince area

4.7 COMPARING COMPLIANCE DURING MOMENTS OF HAND HYGIENE

momentwise[["Moments", "Compliance (%)"]]

# Maximum and Minimum complaince during 5 moments of hand hygiene

max = momentwise.loc[momentwise["Compliance (%)"].idxmax()]
min = momentwise.loc[momentwise["Compliance (%)"].idxmin()]

print(max) # After Patient Contact
print(min) # Before Patient Contact

# The compliance is maximum after getting in contact with patient
# but, the least complaince was reported before patient contact

4.8 DATA COMPARISON DONE!!
