==> notebooks/06_Statistical_Analysis.ipynb <==
This Notebook provides the chisquare and z-test results on the data obtained from the study hand hygiene compliance.

6.1 Importing libraries

import pandas as pd

# for CHI SQUARE TEST
import numpy as np
from scipy.stats import chi2_contingency

# for z-test
from statsmodels.stats.proportion import proportions_ztest

6.3 CHI SQUARE TEST

# CHI SQUARE TEST 1: HCW'S CATEGORY VERSUS COMPLIANCE 
"""  
CATEGORY  Opportunities  ACTION_5-MOMENT    NO ACTION 
Doctor      236            64                 172
Nurses      509            246                263   
GDA/HK      172            73                 99
Others      281            100                181
"""

# Contingency table
# np array - creates a matrix
moment_5 = np.array([
    [64, 172], # Doctors
    [246, 263], # Nurses
    [73, 99], # GDA/HK
    [100, 181] # Other
])

# run a chi square test on this table to get chi2 statistic, probablity, degree of freedom, expected frequencies all in one step
chi2, p, dof, expected = chi2_contingency(moment_5)

'''  
1. chi2 = chi-square statistic
2. p = probability value
3. dof = degree of freedom
4. expected = expected counts/frequency 
'''
print("Chi-square value: ", chi2)
print("Degree of freedom: ", dof)
print("p-value: ", p)


# The compliance behaviour differs widely among hcw's
# nurses showed higher actions while doctors performed the least actions. GDA/HK's compliance was moderate
# targeted training interventions may be required for specific categories of HCW's

# CHI-SQUARE TEST 2: HAND RUB VERSUS HAND WASH PREFERENCES AMONG HCW'S
"""
PROFESSION     ACTION_5_MOMENTS   HANDRUB  HANDWASH
Doctors             64              52       11
Nurses              246             178      67
GDA/HK              73              37       40
Other               100             63       37
"""

preference = np.array([
    [52, 11], # Doctor 
    [178, 67],  # Nurses 
    [37, 40],  # Nurses
    [63, 37]  # Nurses
])


chi2, p, dof, expected = chi2_contingency(preference)

print("Chi-square value: ", chi2)
print("Degree of freedom: ", dof)
print("p-value: ", p)

6.4 Z- TEST

# Z TEST 1 :  CRITICAL AND SEMI CRITICAL COMPARISON

"""
CATEGORY       OPPORTUNITIES   ACTION    
Critical          527           229        
Semi Critical     661           244          
"""
action = np.array([229, 244]) # action- critical, semicritical
opportunities = np.array([527, 661]) # opportunities - critical, semicritical

z_stat, p_val = proportions_ztest(action, opportunities)

print("Z-Statistics: ", z_stat)
print("p-value: ", p_val)




# Z TEST 2: DOCTORS V/S NURSES COMPLIANCE 
"""  
CATEGORY    OPPORTUNITIES   ACTION    
Doctor          236          64         
Nurses          509          246          
"""
action = np.array([64, 246])
opportunities = np.array([236, 509])

z_stat, p_val = proportions_ztest(action, opportunities)

print("Z-Statistics: ", z_stat)
print("p-value: ", p_val)

# the negative value shows that doctors have lower compliance than nurses
# the probability of having equal compiance is observed to be almost zero

# Z TEST 3: GDA/HK AND OTHER COMPLIANCE 

"""  
CATEGORY    OPPORTUNITIES   ACTION      
GDA/HK          172          73         
Others          281          100        
"""

action = np.array([73, 100])
opportunities = np.array([172, 281])

z_stat, p_val = proportions_ztest(action, opportunities)

print("Z-Statistics: ", z_stat)
print("p-value: ", p_val)

# the difference is not statistically significant - the compliance between GDA/HK and Other is roughly the same
