==> notebooks/07_Summary _and_Interpretation.ipynb <==
INTERPRETATION AND SUMMARY OF HAND HYGIENE COMPLIANCE 

This Notebook evaluates the Hand Hygiene compliance among health care workers in different aspects such as:
1. Overall Compliance (5 Moments and 7 Steps)
2. Performance of HCW's in Different Categories (Doctor, Nurse, GDA/HK, Other)
3. Preference of HCW'S for Cleaning hands (Handrub / Handwash)
4. Compliance in Critical and Semi-critical care areas
5. Areawise Compliance among HCW'S
6. Moment wise Compliance
7. visual and Statistical comparison

The main aim of this Study is to analyse the Compliance and identify the Highest and lowest performing Groups and vulnerable areas to understand the patterns and control the infection risk in a tertiary care hospital

7.1 DESCRIPTIVE ANALYSIS

I. OVERALL HAND HYGIENE COMPLIANCE



'''
Total Opportunities = 1198 
Total Actions during 5 moments = 483
Total Actions with 7 steps = 303
Overall Compliance (5 moments) = 40.32 %
Compliance (7 Steps) = 62.73 %
'''


INTERPRETATION

1. The Hand Hygiene compliance reported to be less than half of the total opportunities occured which interprates the unsatisfactory hand hygiene practices among health care workers.
2. The higher Compliance for 7 Steps indicates the HCW's carry out hand hygiene correctly when they choose to do it but they often miss the appropriate indication.

 II. CATEGORY WISE COMPLIANCE


'''
Compliance during 5 Moments

Profession   Compliance (%)
1. Doctors    27.12
2. Nurses     48.33
3. GDA/HK     42.44 
4. Others     35.59 

Compliance with 7 steps

Profession   Compliance (%)
1. Doctors    56.25 
2. Nurses     68.29
3. GDA/HK     61.64 
4. Others     54.00 
'''


INTERPRETATION 

1. Nurses showed the highest compliance in both the categories while doctors performed the least action (particularly in 5 Moments). 
2. GDA/HK indicates that they mostly follow 7 steps when they do hand cleaning during 5 moments of hand hygiene.


III. HAND RUB V/S HAND WASH PREFERENCES


"""
HCW         HR (%)  HW (%)
overall     69.15   32.51
Doctors     81.25   17.19
Nurses      72.36   27.24
GDA/HK      50.68   54.79
Other       63.00   37.00
"""

INTERPRETATION

1. Overall Hand rub is the most preffered choice among HCW's except GDA/HK, as they rely more on Hand wash
2. Doctors prefers Hand rub the most.

IV. CRITICAL AND SEMICRITICAL AREAS COMPLIANCE

''' 
HCW          Critical (%)  Semi-Critical (%)
Overall       43.45          36.91
Doctors       28.07          26.23
Nurses        51.65          43.19
GDA/HK        53.95          33.33
Other         32.63          37.10
'''

INTERPRETATION

1. The compliance in Semi Critical care units drops.
2. Nurses showed maximum compliance in both critical and semicritical areas. 
3. Whereas, the complaince among Doctors reports out to be the least in both the areas


V. Area-Wise Compliance

'''
Areas	 Overall(%)  Doctors(%)  Nurses(%)	GDA/HK(%)  Other (%)
MICU1	  50.98	        43.85	    54.87	 68.57	    33.33
MICU2	  47.30	        16.60	    65.90	 45.45	    35.00
NeuroICU  41.18	        5.80	    57.50	 50.00	    34.48
CCU	      30.47	        13.60	    36.84	 28.57	    25.00
HDUN	  46.38	        20.00	    70.00	 50.00	    52.63
CS	      34.28	        100.00	    14.28	 50.00	    33.33
SR	      40.66	        23.52	    50.00	 44.44	    34.78
EMR	      24.62	        25.00	    30.00	 22.22	    15.90
CR	      40.37	        25.00	    50.00	 23.07	    32.69
Dialysis  45.94	        0.00	     0.00	 5.88	    61.53
'''

###### INTERPRETATION #########

'''
Best-performing units
1. MICU1 - 50.98 %
2. MICU2 - 47.30 %
3. HDUN - 46.38 %
4. Dialysis - 45.94 %

Least-performing Units
1. EMR - 24.62%
2. CCU - 30.47 %
'''

INTERPRETATIONS 
1. Doctors show extremely low compliance in NeuroICU while the highest compliance in CS
2. Nurses reports to be the most compliant in HDUN
3. GDA/HK excel in MICU1

VI. Moment Wise Compliance

'''
Moments                        Compliance (%)
Before Patient Contact              30.13
Before Asceptic Procedure           34.09
After Body Fluid Exposure           32.00
After Patient Contact               53.23
After Touching Patient Surroundings 38.18
 '''

 ###### INTERPRETATION #########
'''
Highest compliance - After Patient Contact
Lowest Compliance - Befor Patient Contact
'''

INTERPRETATIONS

 1. HCW's observed to be more compliant after touching the patient rather than before touching the patient
 2. This indicate poor preventive behaviour

7.2 COMPARATIVE AND VISUAL ANALYSIS (HIGHEST/LOWEST VALUES)

'''
*******DOMAIN WISE COMPLIANCE COMPARISON********
DOMAIN              HIGHEST       LOWEST

5 Moments           Nurses       Doctors
7 Steps             Nurses       Others
Hand Rub            Doctors      GDA/HK
Hand wash           GDA/HK       DOCTORS
Critical            GDA/HK       DOCTORS
Semi Critical       Nurses       DOCTORS 
Area wise           MICU1        EMR
Moment wise         Aft.Pt.      Bf.Pt.
'''

''' 
*******AREA WISE COMPLIANCE COMPARISON********

AREA             HIGHEST       LOWEST

MICU1            GDA/HK        Other
MICU2            Nurses        Doctors
NeuroICU         Nurses        Doctors
CCU              Nurses        Doctors
HDUN             Nurses        Doctors
CS               Doctors       Nurses
SR               Nurses        Doctors
EMR              Nurses        Other
CR               Nurses        GDA/HK
Dialysis         Other         GDA/HK
'''

7.3 STATISTICAL ANALYSIS

I. HCW AND 5 MOMENTS COMPLIANCE (Chi Square)

'''
Chi-square value:  33.60246465606425
Degree of freedom:  3
p-value:  2.4034472645234014e-07
'''

INTERPRETATION

1. non uniform compliance among HCW's
2. The difference is not likely to be due to a chance, it indicates real behavioural variation among health care worker categories

II. HAND RUB VS HAND WASH PREFERENCES AMONG HCW'S (Chi square)

'''
Chi-square value:  23.804044344536592
Degree of freedom:  3
p-value:  2.7447072779652652e-05
'''


Interpretation

1. There is variation in pattern of preferences among HCW's
2. It does not show any random fluctuation

III. CRITICAL AND SEMICRITICAL COMPLIANCE (Z-Test)

'''Z-Statistics:  2.2876594584859755
p-value:  0.022157360388318285'''

INTERPRETATIONS

1. Compliance portions between critical and semicritical areas showed significant difference.
2. Low compliance in Semicritical areas suggest risk in behaviour, where as HCW are more compliant in critical areas

IV. DOCTORS VS NURSES COMPLIANCE (Z-Test)

'''Z-Statistics:  -5.464333431842507
p-value:  4.646494430309193e-08'''

INTERPRETATION

1. Highly significant difference between these 2 groups
2. Nurses showed higher compliance in comparison to doctors

V. COMPLIANCE IN GDA/HK AND OTHER (Z-Test)

'''Z-Statistics:  1.4573071183111332
p-value:  0.14503162278541362'''

INTERPRETATION

1. No significant difference between 2 groups
2. The compliance among both groups are statistically comparable and shows similar behaviour and training patterns

7.4 SUMMARY

1. Overall, the data indicates hand hygiene behaviours among healthcare workers in different hospital areas
2. Nurses demonstrates the consistant superior compliance in comparison to doctors.
3. The preference in method of hand cleaning also varies among HCW's
4. Compliance is higher in Critical care areas as compared to Semicritical care units
5. The data highlights the need for targeted training interventions, particularly for doctors and in low performing areas.
